requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.1.0
autopep8==1.5.4
beautifulsoup4==4.9.1
colorama==0.4.3
cycler==0.10.0
dj-static==0.0.6
Django==3.1.4
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-rest-framework==0.1.0
django-tables2==2.3.3
djangorestframework==3.12.2
docopt==0.6.2
html-linter==0.4.0
iniconfig==1.0.1
isort==5.4.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
numpy==1.19.4
packaging==20.4
Pillow==7.2.0
plotly==4.14.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
template-remover==0.1.9
toml==0.10.1
wrapt==1.12.1